requirements.txt

flask
gunicorn
psycopg2-binary
werkzeug
reportlab
pandas
dropbox
